Backend/python backend/requirements.txt
Flask>=2.0
numpy
opencv-python
Pillow
easyocr
face_recognition
scikit-image

